native: switched to f2c defined integers

+++ src/NativeWrappers/GotoBlas2/f2c.h
@@ -7,8 +7,7 @@
 #ifndef F2C_INCLUDE
 #define F2C_INCLUDE
 
-typedef int integer;
-//typedef long int integer;
+typedef long int integer;
 
 typedef unsigned long int uinteger;
 typedef char *address;
